% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add-ecoregions.R
\name{add_ecoregions}
\alias{add_ecoregions}
\title{Join the three tables of WildFinder}
\usage{
add_ecoregions(x, y, z)
}
\arguments{
\item{x}{a \code{tibble} (or \code{data.frame}). The output of \code{\link[=read_sp_list]{read_sp_list()}}.}

\item{y}{a \code{tibble} (or \code{data.frame}). The output of \code{\link[=read_sp_eco]{read_sp_eco()}}.}

\item{z}{a \code{tibble} (or \code{data.frame}). The output of \code{\link[=read_eco_list]{read_eco_list()}}.}
}
\value{
A \code{tibble} with the following two columns: \code{sci_name} and
\code{ecoregion}.
}
\description{
This function adds the ecoregions information to the species list by joining
the three WildFinder tables.
}
\examples{
\dontrun{
## Download data ----
dl_wildfinder_data()

## Import data ----
sp_list  <- read_sp_list(here::here("data", "wildfinder", 
                                    "wildfinder-mammals_list.csv"))
sp_eco   <- read_sp_eco(here::here("data", "wildfinder", 
                                   "wildfinder-ecoregions_species.csv"))
eco_list <- read_eco_list(here::here("data", "wildfinder", 
                                     "wildfinder-ecoregions_list.csv"))
                                   
## Susbet species ----
sp_list <- select_species(sp_list, family = "Ursidae")

## Add ecoregions ----
sp_eco_list <- add_ecoregions(sp_list, sp_eco, eco_list)
}
}
