% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count-ecoregions.R
\name{count_ecoregions}
\alias{count_ecoregions}
\title{Count ecoregions by species}
\usage{
count_ecoregions(x)
}
\arguments{
\item{x}{a \code{tibble} (or \code{data.frame}). The output of \code{\link[=add_ecoregions]{add_ecoregions()}}.}
}
\value{
A \code{tibble} with the following two columns: \code{sci_name} and
\code{n_ecoregions}.
}
\description{
This function counts the number of distinct ecoregions per species.
}
\examples{
\dontrun{
## Download data ----
dl_wildfinder_data()

## Import data ----
sp_list  <- read_sp_list(here::here("data", "wildfinder", 
                                    "wildfinder-mammals_list.csv"))
sp_eco   <- read_sp_eco(here::here("data", "wildfinder", 
                                   "wildfinder-ecoregions_species.csv"))
eco_list <- read_eco_list(here::here("data", "wildfinder", 
                                     "wildfinder-ecoregions_list.csv"))
                                   
## Susbet species ----
sp_list <- select_species(sp_list, family = "Ursidae")

## Add ecoregions ----
sp_eco_list <- add_ecoregions(sp_list, sp_eco, eco_list)

## Count ecoregions ----
count_ecoregions(sp_eco_list)
}
}
