% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-counts.R
\name{plot_counts}
\alias{plot_counts}
\title{Barplot of the number of ecoregions by species}
\usage{
plot_counts(x)
}
\arguments{
\item{x}{a \code{tibble} (or \code{data.frame}). The output of \code{\link[=count_ecoregions]{count_ecoregions()}}.}
}
\value{
A \code{ggplot} object.
}
\description{
This function plots a barplot of the number of distinct ecoregions per
species using the package \code{\link{ggplot2}}.
}
\examples{
\dontrun{
## Download data ----
dl_wildfinder_data()

## Import data ----
sp_list  <- read_sp_list(here::here("data", "wildfinder", 
                                    "wildfinder-mammals_list.csv"))
sp_eco   <- read_sp_eco(here::here("data", "wildfinder", 
                                   "wildfinder-ecoregions_species.csv"))
eco_list <- read_eco_list(here::here("data", "wildfinder", 
                                     "wildfinder-ecoregions_list.csv"))
                                   
## Susbet species ----
sp_list <- select_species(sp_list, family = "Ursidae")

## Add ecoregions ----
sp_eco_list <- add_ecoregions(sp_list, sp_eco, eco_list)

## Count ecoregions ----
counts_data <- count_ecoregions(sp_eco_list)

## Plot ----
plot_counts(counts_data)
}
}
