% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read-files.R
\name{read_eco_list}
\alias{read_eco_list}
\title{Read ecoregions list file - WWF WildFinder}
\usage{
read_eco_list(file)
}
\arguments{
\item{file}{a character of length 1. The path to the csv file.}
}
\value{
A \code{tibble} with 798 rows (ecoregions) and 4 variables describing
ecoregions.
}
\description{
This function reads the WWF WildFinder \code{csv} file named
\code{wildfinder-ecoregions_list.csv} stored in \verb{data/wildfinder/}.

\strong{Note:} the function \code{\link[=dl_wildfinder_data]{dl_wildfinder_data()}} must have been run first.
}
\examples{
\dontrun{
## Download data ----
dl_wildfinder_data()

## Import data ----
eco_list <- read_eco_list(here::here("data", "wildfinder", 
                                     "wildfinder-ecoregions_list.csv"))
}
}
