% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read-files.R
\name{read_sp_eco}
\alias{read_sp_eco}
\title{Read ecoregions-species file - WWF WildFinder}
\usage{
read_sp_eco(file)
}
\arguments{
\item{file}{a character of length 1. The path to the csv file.}
}
\value{
A \code{tibble} with 350,045 rows (ecoregion-species) and 2 variables.
This table is used to merge the two other \code{tibble}.
}
\description{
This function reads the WWF WildFinder \code{csv} file named
\code{wildfinder-ecoregions_species.csv} stored in \verb{data/wildfinder/}.

\strong{Note:} the function \code{\link[=dl_wildfinder_data]{dl_wildfinder_data()}} must have been run first.
}
\examples{
\dontrun{
## Download data ----
dl_wildfinder_data()

## Import data ----
sp_eco <- read_sp_eco(here::here("data", "wildfinder", 
                                 "wildfinder-ecoregions_species.csv"))
}
}
