% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read-files.R
\name{read_sp_list}
\alias{read_sp_list}
\title{Read species list file - WWF WildFinder}
\usage{
read_sp_list(file)
}
\arguments{
\item{file}{a character of length 1. The path to the csv file.}
}
\value{
A \code{tibble} with 4,936 rows (species) and 8 variables describing
species.
}
\description{
This function reads the WWF WildFinder \code{csv} file named
\code{wildfinder-mammals_list.csv} stored in \verb{data/wildfinder/}.

\strong{Note:} the function \code{\link[=dl_wildfinder_data]{dl_wildfinder_data()}} must have been run first.
}
\examples{
\dontrun{
## Download data ----
dl_wildfinder_data()

## Import data ----
sp_list <- read_sp_list(here::here("data", "wildfinder", 
                                   "wildfinder-mammals_list.csv"))
}
}
