% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select-species.R
\name{select_species}
\alias{select_species}
\title{Select all species of a Mammal family}
\usage{
select_species(x, family = NULL)
}
\arguments{
\item{x}{a \code{tibble} (or \code{data.frame}). The output of \code{\link[=read_sp_list]{read_sp_list()}}.}

\item{family}{a \code{character}. The name of a Mammal family (or several) to
subset.}
}
\value{
A \code{tibble} with the following three columns: \code{species_id},
\code{sci_name}, and \code{family}.
}
\description{
This function selects all the species of a Mammal family (argument \code{family}).
\code{wildfinder-mammals_list.csv} stored in \verb{data/wildfinder/}.

\strong{Note:} the function \code{\link[=dl_wildfinder_data]{dl_wildfinder_data()}} must have been run first.
}
\examples{
\dontrun{
## Download data ----
dl_wildfinder_data()

## Import data ----
sp_list <- read_sp_list(here::here("data", "wildfinder", 
                                   "wildfinder-mammals_list.csv"))
                                   
## Susbet species ----
select_species(sp_list, family = "Ursidae")
}
}
